requirements.txt
about-time==4.2.1
alive-progress==3.1.5
asttokens==2.4.1
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
configobj==5.0.8
configparser==7.0.0
cryptography==42.0.7
debugpy==1.8.1
decorator==5.1.1
et-xmlfile==1.1.0
etelemetry==0.3.1
executing==2.0.1
filelock==3.14.0
fitz==0.0.1.dev2
frozendict==2.4.4
grapheme==0.6.0
html5lib==1.1
httplib2==0.22.0
idna==3.7
ipykernel==6.29.4
ipython==8.24.0
isodate==0.6.1
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
looseversion==1.3.0
lxml==5.2.2
matplotlib-inline==0.1.7
multitasking==0.0.11
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
parso==0.8.4
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.0
peewee==3.17.5
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.45
prov==2.0.0
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
pydot==2.0.0
Pygments==2.18.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pyxnat==1.6.2
pyzmq==26.0.3
rdflib==7.0.0
requests==2.32.2
scipy==1.13.1
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
traits==6.3.2
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
yfinance==0.2.40
